+++ cli/cmd/utils/utils.go
@@ -53,6 +53,7 @@ import (
 	opv1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
 	sfv1 "github.com/softwarefactory-project/sf-operator/api/v1"
 	controllers "github.com/softwarefactory-project/sf-operator/controllers"
+	ctrlutils "github.com/softwarefactory-project/sf-operator/controllers/libs/utils"
 
 	"k8s.io/client-go/kubernetes"
 )
@@ -137,21 +138,6 @@ func SetLogger(command *cobra.Command) {
 	ctrl.SetLogger(zap.New(zap.UseFlagOptions(&opts)))
 }
 
-// logI logs a message with the INFO log Level
-func logI(msg string) {
-	ctrl.Log.Info(msg)
-}
-
-// logI logs a message with the DEBUG log Level
-func logD(msg string) {
-	ctrl.Log.V(1).Info(msg)
-}
-
-// logI logs a message with the Error log Level
-func logE(err error, msg string) {
-	ctrl.Log.Error(err, msg)
-}
-
 func GetCLIContext(command *cobra.Command) (SoftwareFactoryConfigContext, error) {
 
 	// This is usually called for every CLI command so here let's set the Logger settings
@@ -164,10 +150,10 @@ func GetCLIContext(command *cobra.Command) (SoftwareFactoryConfigContext, error)
 	if configPath != "" {
 		ctxName, cliContext, err = getContextFromFile(command)
 		if err != nil {
-			logE(err, "Could not load config file")
+			ctrlutils.LogE(err, "Could not load config file")
 			os.Exit(1)
 		} else {
-			logD("Using configuration context " + ctxName)
+			ctrlutils.LogD("Using configuration context " + ctxName)
 		}
 	}
 	// Override with defaults
@@ -190,12 +176,12 @@ func GetCLIContext(command *cobra.Command) (SoftwareFactoryConfigContext, error)
 	if cliContext.Dev.SFOperatorRepositoryPath == "" {
 		defaultSFOperatorRepositoryPath, getwdErr := os.Getwd()
 		if getwdErr != nil {
-			logE(getwdErr,
+			ctrlutils.LogE(getwdErr,
 				"sf-operator-repository-path is not set in `dev` section of the configuration file and unable to determine the current working directory")
 			os.Exit(1)
 		}
 		cliContext.Dev.SFOperatorRepositoryPath = defaultSFOperatorRepositoryPath
-		logD("Using current working directory for sf-operator-repository-path: " + cliContext.Dev.SFOperatorRepositoryPath)
+		ctrlutils.LogD("Using current working directory for sf-operator-repository-path: " + cliContext.Dev.SFOperatorRepositoryPath)
 	}
 	return cliContext, nil
 }
@@ -239,7 +225,7 @@ func CreateKubernetesClient(contextName string) (client.Client, error) {
 func CreateKubernetesClientOrDie(contextName string) client.Client {
 	cli, err := CreateKubernetesClient(contextName)
 	if err != nil {
-		ctrl.Log.Error(err, "Error creating Kubernetes client")
+		ctrlutils.LogE(err, "Error creating Kubernetes client")
 		os.Exit(1)
 	}
 	return cli
@@ -249,7 +235,7 @@ func GetCLIENV(kmd *cobra.Command) (string, ENV) {
 
 	cliCtx, err := GetCLIContext(kmd)
 	if err != nil {
-		ctrl.Log.Error(err, "Error initializing CLI:")
+		ctrlutils.LogE(err, "Error initializing CLI:")
 		os.Exit(1)
 	}
 
@@ -283,7 +269,7 @@ func DeleteOrDie(env *ENV, obj client.Object, opts ...client.DeleteOption) bool 
 		return false
 	} else if err != nil {
 		msg := fmt.Sprintf("Error while deleting %s \"%s\"", reflect.TypeOf(obj).Name(), obj.GetName())
-		ctrl.Log.Error(err, msg)
+		ctrlutils.LogE(err, msg)
 		os.Exit(1)
 	}
 	return true
@@ -295,7 +281,7 @@ func GetMOrDie(env *ENV, name string, obj client.Object) bool {
 		return false
 	} else if err != nil {
 		msg := fmt.Sprintf("Error while fetching %s \"%s\"", reflect.TypeOf(obj).Name(), name)
-		ctrl.Log.Error(err, msg)
+		ctrlutils.LogE(err, msg)
 		os.Exit(1)
 	}
 	return true
@@ -303,33 +289,33 @@ func GetMOrDie(env *ENV, name string, obj client.Object) bool {
 
 func UpdateROrDie(env *ENV, obj client.Object) {
 	var msg = fmt.Sprintf("Updating %s \"%s\" in %s", reflect.TypeOf(obj).Name(), obj.GetName(), env.Ns)
-	ctrl.Log.Info(msg)
+	ctrlutils.LogI(msg)
 	if err := env.Cli.Update(env.Ctx, obj); err != nil {
 		msg = fmt.Sprintf("Error while updating %s \"%s\"", reflect.TypeOf(obj).Name(), obj.GetName())
-		ctrl.Log.Error(err, msg)
+		ctrlutils.LogE(err, msg)
 		os.Exit(1)
 	}
 	msg = fmt.Sprintf("%s \"%s\" updated", reflect.TypeOf(obj).Name(), obj.GetName())
-	ctrl.Log.Info(msg)
+	ctrlutils.LogI(msg)
 }
 
 func CreateROrDie(env *ENV, obj client.Object) {
 	var msg = fmt.Sprintf("Creating %s \"%s\" in %s", reflect.TypeOf(obj).Name(), obj.GetName(), env.Ns)
-	ctrl.Log.Info(msg)
+	ctrlutils.LogI(msg)
 	obj.SetNamespace(env.Ns)
 	if err := env.Cli.Create(env.Ctx, obj); err != nil {
 		msg = fmt.Sprintf("Error while creating %s \"%s\"", reflect.TypeOf(obj).Name(), obj.GetName())
-		ctrl.Log.Error(err, msg)
+		ctrlutils.LogE(err, msg)
 		os.Exit(1)
 	}
 	msg = fmt.Sprintf("%s \"%s\" created", reflect.TypeOf(obj).Name(), obj.GetName())
-	ctrl.Log.Info(msg)
+	ctrlutils.LogI(msg)
 }
 
 func DeleteAllOfOrDie(env *ENV, obj client.Object, opts ...client.DeleteAllOfOption) {
 	if err := env.Cli.DeleteAllOf(env.Ctx, obj, opts...); err != nil {
 		var msg = "Error while deleting"
-		ctrl.Log.Error(err, msg)
+		ctrlutils.LogE(err, msg)
 		os.Exit(1)
 	}
 }
@@ -337,7 +323,7 @@ func DeleteAllOfOrDie(env *ENV, obj client.Object, opts ...client.DeleteAllOfOpt
 func GetCLIctxOrDie(kmd *cobra.Command, args []string, allowedArgs []string) SoftwareFactoryConfigContext {
 	cliCtx, err := GetCLIContext(kmd)
 	if err != nil {
-		ctrl.Log.Error(err, "Error initializing:")
+		ctrlutils.LogE(err, "Error initializing:")
 		os.Exit(1)
 	}
 	if len(allowedArgs) == 0 {
@@ -346,7 +332,7 @@ func GetCLIctxOrDie(kmd *cobra.Command, args []string, allowedArgs []string) Sof
 	} else {
 		argumentError := errors.New("argument must be in: " + strings.Join(allowedArgs, ", "))
 		if len(args) != 1 {
-			ctrl.Log.Error(argumentError, "Need one argument")
+			ctrlutils.LogE(argumentError, "Need one argument")
 			os.Exit(1)
 		}
 		for _, a := range allowedArgs {
@@ -354,7 +340,7 @@ func GetCLIctxOrDie(kmd *cobra.Command, args []string, allowedArgs []string) Sof
 				return cliCtx
 			}
 		}
-		ctrl.Log.Error(argumentError, "Unknown argument "+args[0])
+		ctrlutils.LogE(argumentError, "Unknown argument "+args[0])
 		os.Exit(1)
 	}
 	return SoftwareFactoryConfigContext{}
@@ -380,8 +366,8 @@ func RunCmdWithEnvOrDie(environ []string, cmd string, args ...string) string {
 	kmd.Env = append(os.Environ(), environ...)
 	out, err := kmd.CombinedOutput()
 	if err != nil {
-		logE(err, "Could not run command '"+cmd+"'")
-		logI("Captured output:\n" + string(out))
+		ctrlutils.LogE(err, "Could not run command '"+cmd+"'")
+		ctrlutils.LogI("Captured output:\n" + string(out))
 		os.Exit(1)
 	}
 	return string(out)
@@ -397,7 +383,7 @@ func EnsureNamespaceOrDie(env *ENV, name string) {
 		ns.Name = name
 		CreateROrDie(env, &ns)
 	} else if err != nil {
-		ctrl.Log.Error(err, "Error checking namespace "+name)
+		ctrlutils.LogE(err, "Error checking namespace "+name)
 		os.Exit(1)
 	}
 }
@@ -412,7 +398,7 @@ func EnsureServiceAccountOrDie(env *ENV, name string) {
 func WriteContentToFile(filePath string, content []byte, mode fs.FileMode) {
 	err := os.WriteFile(filePath, content, mode)
 	if err != nil {
-		ctrl.Log.Error(err, "Can not write a file "+filePath)
+		ctrlutils.LogE(err, "Can not write a file "+filePath)
 		os.Exit(1)
 	}
 }
@@ -425,7 +411,7 @@ func VarListToMap(varsList []string) map[string]string {
 		tokens := strings.Split(v, "=")
 
 		if len(tokens) != 2 {
-			ctrl.Log.Error(errors.New("parse error"), "parsed value `"+v+"` needs to be defined as 'foo=bar'")
+			ctrlutils.LogE(errors.New("parse error"), "parsed value `"+v+"` needs to be defined as 'foo=bar'")
 			os.Exit(1)
 		}
 		vars[tokens[0]] = tokens[1]
@@ -436,7 +422,7 @@ func VarListToMap(varsList []string) map[string]string {
 func CreateDirectory(dirPath string, mode fs.FileMode) {
 	err := os.MkdirAll(dirPath, mode)
 	if err != nil {
-		ctrl.Log.Error(err, "Can not create directory "+dirPath)
+		ctrlutils.LogE(err, "Can not create directory "+dirPath)
 		os.Exit(1)
 	}
 }
@@ -454,7 +440,7 @@ func GetClientset(kubeContext string) (*rest.Config, *kubernetes.Clientset) {
 	restConfig := controllers.GetConfigContextOrDie(kubeContext)
 	kubeClientset, err := kubernetes.NewForConfig(restConfig)
 	if err != nil {
-		ctrl.Log.Error(err, "Could not instantiate Clientset")
+		ctrlutils.LogE(err, "Could not instantiate Clientset")
 		os.Exit(1)
 	}
 	return restConfig, kubeClientset
@@ -482,7 +468,7 @@ func RunRemoteCmd(kubeContext string, namespace string, podName string, containe
 	if err != nil {
 		errMsg := fmt.Sprintf("Command \"%s\" [Pod: %s - Container: %s] failed with the following stderr: %s",
 			strings.Join(cmdArgs, " "), podName, containerName, errorBuffer.String())
-		ctrl.Log.Error(err, errMsg)
+		ctrlutils.LogE(err, errMsg)
 		os.Exit(1)
 	}
 	return buffer
@@ -493,10 +479,10 @@ func ReadYAMLToMapOrDie(filePath string) map[string]interface{} {
 	secretContent := make(map[string]interface{})
 	err := yaml.Unmarshal(readFile, &secretContent)
 	if err != nil {
-		ctrl.Log.Error(err, "Problem on reading the file content")
+		ctrlutils.LogE(err, "Problem on reading the file content")
 	}
 	if len(secretContent) == 0 {
-		ctrl.Log.Error(errors.New("file is empty"), "The file is empty or it does not exist!")
+		ctrlutils.LogE(errors.New("file is empty"), "The file is empty or it does not exist!")
 		os.Exit(1)
 	}
 	return secretContent
@@ -505,7 +491,7 @@ func ReadYAMLToMapOrDie(filePath string) map[string]interface{} {
 func GetKubectlPath() string {
 	kubectlPath, err := exec.LookPath("kubectl")
 	if err != nil {
-		ctrl.Log.Error(errors.New("no kubectl binary"),
+		ctrlutils.LogE(errors.New("no kubectl binary"),
 			"No 'kubectl' binary found. Please install the 'kubectl' binary before attempting a restore")
 		os.Exit(1)
 	}
@@ -519,7 +505,7 @@ func ExecuteKubectlClient(ns string, podName string, containerName string, execu
 
 	err := cmd.Run()
 	if err != nil {
-		ctrl.Log.Error(err, "There is an issue on executing command: "+executeCommand)
+		ctrlutils.LogE(err, "There is an issue on executing command: "+executeCommand)
 		os.Exit(1)
 	}