+++ controllers/nodepool.go
@@ -341,7 +341,7 @@ func (r *SFController) ensureNodepoolPromRule(cloudsYaml map[string]interface{})
 		return false
 	} else {
 		if !utils.MapEquals(&currentPromRule.ObjectMeta.Annotations, &annotations) {
-			r.log.V(1).Info("Nodepool default Prometheus rules changed, updating...")
+			utils.LogI("Nodepool default Prometheus rules changed, updating...")
 			currentPromRule.Spec = desiredNodepoolPromRule.Spec
 			currentPromRule.ObjectMeta.Annotations = annotations
 			r.UpdateR(&currentPromRule)
@@ -790,7 +790,7 @@ func (r *SFController) DeployNodepoolLauncher(statsdExporterVolume apiv1.Volume,
 	current := appsv1.Deployment{}
 	if r.GetM(LauncherIdent, &current) {
 		if !utils.MapEquals(&current.Spec.Template.ObjectMeta.Annotations, &annotations) {
-			r.log.V(1).Info("Nodepool-launcher configuration changed, rollout pods ...")
+			utils.LogI("Nodepool-launcher configuration changed, rollout pods ...")
 			current.Spec = nl.DeepCopy().Spec
 			r.UpdateR(&current)
 			return false